refactor: exception
