+++ history-service/src/main/java/com/study/history/service/repository/HistoryRepository.java
@@ -1,5 +1,6 @@
-package com.study.history.service;
+package com.study.history.service.repository;
 
+import com.study.history.service.domain.History;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
