+++ history-service/src/main/java/com/study/history/service/response/exception/ExceptionResponse.java
@@ -1,0 +1,13 @@
+package com.study.history.service.response.exception;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class ExceptionResponse {
+    private String errorMessage;
+}
