+++ history-service/src/main/java/com/study/history/service/response/exception/RestExceptionHandler.java
@@ -1,0 +1,15 @@
+package com.study.history.service.response.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class RestExceptionHandler {
+    @ExceptionHandler(UserIdException.class)
+    public ResponseEntity<ExceptionResponse> handleUserIdException(UserIdException ex) {
+        ExceptionResponse response = new ExceptionResponse(ex.getMessage());
+        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    }
+}
