+++ history-service/src/main/java/com/study/history/service/response/exception/UserIdException.java
@@ -1,0 +1,8 @@
+package com.study.history.service.response.exception;
+
+public class UserIdException extends RuntimeException {
+
+    public UserIdException(String message) {
+        super(message);
+    }
+}
